@@ -5,16 +5,16 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.4</version>
+		<version>3.2.1</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>org.bondor</groupId>
 	<artifactId>family-dashboard</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<properties>
-		<java.version>17</java.version>
-		<maven.compiler.source>17</maven.compiler.source>
-		<maven.compiler.target>17</maven.compiler.target>
+		<java.version>21</java.version>
+		<maven.compiler.source>21</maven.compiler.source>
+		<maven.compiler.target>21</maven.compiler.target>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -42,7 +42,7 @@
 			<!-- EXIF lib https://github.com/drewnoakes/metadata-extractor/ -->
 		  <groupId>com.drewnoakes</groupId>
 		  <artifactId>metadata-extractor</artifactId>
-		  <version>2.18.0</version>
+		  <version>2.19.0</version>
 		</dependency>
 
 	</dependencies>
