@@ -10,7 +10,7 @@ import java.io.InputStream;
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
 import org.springframework.core.io.ClassPathResource;
